+++ src/main/java/com/atlauncher/data/minecraft/loaders/forge/ForgeLoader.java
@@ -209,4 +209,9 @@ public class ForgeLoader implements Loader {
     public List<Library> getInstallLibraries() {
         return null;
     }
+
+    @Override
+    public LoaderVersion getLoaderVersion() {
+        return new LoaderVersion(version, rawVersion, false, "Forge");
+    }
 }