+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 3.3.1.0
 
-- Install packs from Curse through dragging and dropping the Install button
-- Sdd in servers tab to manage servers created through the launcher
+- Add in servers tab to manage servers created through the launcher
+- Add Curse packs by dragging the pack zip in instances tab